backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
langchain
langchain-community
langchain-openai
langchain-google-genai
duckduckgo-search
python-dotenv
httpx
numpy
sqlmodel

passlib[bcrypt]
python-jose[cryptography]
python-multipart
